Guided by https://www.youtube.com/watch?v=DhUpxWjOhME Large refactor of testing: - moved src to src folder - moved tests to tests folder
Src configured to build a package which is then used by /tests to remove dependency on the source code within the same directory

Added several config files:
-setup.py (boilerplate)
-setup.cfg - contains package metadata and package config
-basic tox config for test automation
-pyproject.toml - configures test packages - directories they should target and config for how they should run

-added flake8 (style enforcement)
-added mypy (type hinting)
