+++ requirements_dev.txt
@@ -1,0 +1,6 @@
+flake8==5.0.4
+tox==3.25.1
+pytest==7.1.2
+pytest-cov==3.0.0
+mypy==0.971
+
