+++ seutp.cfg
@@ -1,0 +1,32 @@
+[metadata]
+name = ReplaysLolRedditBot
+description = RedditBot to scrape comments containing LoL statistics website links
+author = George Thomas
+classisfiers = Programming Language :: Python :: 3.10
+
+[options]
+packages =
+    ReplaysLolRedditBot
+install_requires =
+    beautifulsoup4==4.11.1
+    praw==7.6.0
+    python-dotenv==0.20.0
+    pytest~=7.1.2
+python_requires = >=3.10
+package_dir =
+    =src
+zip_safe = no
+
+[options.extras_require]
+testing =
+    flake8>=5.0.4
+    tox>=3.25.1
+    pytest>=7.1.2
+    pytest-cov>=3.0.0
+    mypy>=0.971
+
+[options.package_data]
+ReplaysLolRedditBot = py.typed
+
+[flake8]
+max-line-length = 160
