+++ src/ReplaysLolRedditBot/Errors.py
