+++ src/ReplaysLolRedditBot/ReplaysLolRedditBot.py
@@ -5,8 +5,8 @@ from time import sleep
 from bs4 import BeautifulSoup
 from dotenv import load_dotenv
 from Errors import SubmissionExistsException
-from RedditBotHelpers import *
-import Errors
+from Helpers import *
+
 
 def main():
     load_dotenv()
@@ -21,9 +21,9 @@ def main():
         password=os.environ.get('REDDIT_PASSWORD')
     )
 
-    path = "/Comments.json"
+    path = "../../Comments.json"
     if os.path.exists(path):
-        with open("Comments.json", "r") as jsonfile:
+        with open("../../Comments.json", "r") as jsonfile:
             scraped_submissions = json.load(jsonfile)
     else:
         scraped_submissions = []
@@ -35,13 +35,13 @@ def main():
                     for link in soup.findAll('a', href=True):
                         if get_matches_from_link(link.attrs['href']):
                             if is_submission_id_present_in_list_of_dictionaries(str(submission.id),
-                                                                                                 scraped_submissions):
+                                                                                scraped_submissions):
                                 raise SubmissionExistsException
                             data = {"href": link.attrs["href"], "submission_id": str(submission.id)}
                             print(str(data["href"] + " has not already been scraped, continuing"))
                             scraped_submissions.append(data)
                             print("data appended")
-                            with open("Comments.json", "w") as jsonfile:
+                            with open("../../Comments.json", "w") as jsonfile:
                                 json.dump(scraped_submissions, jsonfile, indent=4)
                             print("json dumped")
                             break
@@ -55,6 +55,3 @@ def main():
 
 if __name__ == "__main__":
     main()
-
-# SCOPE CREEP
-# Also raise an error if the submission is > 12 hours
