+++ tox.ini
@@ -1,0 +1,27 @@
+[tox]
+minversion = 3.25.1
+envlist = py310, flake8, mypy
+isolated_build = true
+
+[gh-actions]
+python =
+     3.10: py310, mypy, flake8
+
+[testenv]
+setenv =
+    PYTHONPATH = {toxinidir}
+deps =
+    -r{toxinidir}/requirements_dev.txt
+commands =
+    pytest --basetemp={envtmpdir}
+
+[testenv:flake8]
+basepython = python3.10
+deps = flake8
+commands = flake8 src tests
+
+[testenv:mypy]
+basepython = python3.10
+deps =
+    -r{toxinidir}/requirements_dev.txt
+commands = mypy src
